pipelines/yolov7_segmentation/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
absl-py==2.1.0
    # via tensorboard
annotated-types==0.7.0
    # via pydantic
asttokens==2.4.1
    # via stack-data
beartype==0.17.2
    # via picsellia
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.1.1
    # via ipython
deprecation==2.1.0
    # via picsellia
exceptiongroup==1.2.2
    # via ipython
executing==2.1.0
    # via stack-data
filelock==3.16.1
    # via
    #   torch
    #   triton
fonttools==4.54.1
    # via matplotlib
fsspec==2024.10.0
    # via torch
grpcio==1.67.0
    # via tensorboard
idna==3.10
    # via requests
ipython==8.29.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via torch
kiwisolver==1.4.7
    # via matplotlib
markdown==3.7
    # via tensorboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.9.2
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.7
    # via ipython
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==1.23.5
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   pandas
    #   scipy
    #   seaborn
    #   tensorboard
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
opencv-python==4.8.1.78
    # via -r requirements.in
orjson==3.10.10
    # via picsellia
packaging==24.1
    # via
    #   deprecation
    #   matplotlib
    #   tensorboard
pandas==2.1.0
    # via
    #   -r requirements.in
    #   seaborn
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
picsellia==6.18.3
    # via -r requirements.in
picsellia-annotations==0.9.0
    # via picsellia
picsellia-connexion-services==0.2.1
    # via picsellia
pillow==10.4.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   picsellia
    #   torchvision
prompt-toolkit==3.0.48
    # via ipython
protobuf==3.20.0
    # via
    #   -r requirements.in
    #   tensorboard
psutil==6.1.0
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.9.2
    # via
    #   picsellia
    #   picsellia-annotations
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via ipython
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   picsellia
requests==2.32.3
    # via
    #   -r requirements.in
    #   picsellia
    #   picsellia-connexion-services
scipy==1.14.1
    # via -r requirements.in
seaborn==0.13.2
    # via -r requirements.in
setuptools==75.2.0
    # via tensorboard
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
    #   tensorboard
stack-data==0.6.3
    # via ipython
sympy==1.13.1
    # via torch
tabulate==0.9.0
    # via -r requirements.in
tdqm==0.0.1
    # via picsellia
tensorboard==2.18.0
    # via -r requirements.in
tensorboard-data-server==0.7.2
    # via tensorboard
thop==0.1.1.post2209072238
    # via -r requirements.in
torch==2.5.0
    # via
    #   -r requirements.in
    #   thop
    #   torchvision
torchvision==0.20.0
    # via -r requirements.in
tqdm==4.66.6
    # via
    #   -r requirements.in
    #   tdqm
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   ipython
    #   pydantic
    #   pydantic-core
    #   torch
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.0.6
    # via tensorboard
xmltodict==0.13.0
    # via picsellia-annotations
